# docs/swagger.yaml
basePath: /api/v1
definitions:
  AppServer:
    properties:
      addr:
        type: string
      agent_pwd:
        type: string
      agent_usr:
        type: string
      agent_version:
        type: string
      descr:
        type: string
      name:
        type: string
      port:
        type: string
      properties:
        additionalProperties:
          type: string
        type: object
      version:
        type: string
    type: object
  Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
...
  serialize.UserInfo:
    properties:
      descr:
        example: дополнительное описание
        type: string
      name:
        example: admin
        type: string
      password:
        type: string
      password_auth_allowed:
        type: boolean
      sys_auth_allowed:
        type: boolean
      sys_user_name:
        type: string
    type: object
host: localhost:3001
info:
  contact:
    email: khorevaa@yandex.ru
    name: API Support
  description: KUBOdin Swagger UI
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 'KUBOdin: Remote Administration for 1S.Enterprise Application Servers'
  version: "1.0"
paths:
  /app:
    get:
      consumes:
      - application/json
      description: Получение списка зарегистрированных серверов 1С.Предприятие
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/AppServer'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
      summary: Получение списка зарегистрированных серверов 1С.Предприятие
      tags:
      - app
    post:
      consumes:
      - application/json
      description: выполняет регистрацию сервера 1С.Предприятие в приложении
      parameters:
      - description: app info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/AppServer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/AppServer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: выполняет регистрацию сервера 1С.Предприятие в приложении
      tags:
      - app
  /app/{app}:
    delete:
      consumes:
      - application/json
      description: Удаление информации о регистрации сервера приложений 1С.Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Удаление информации о регистрации сервера приложений 1С.Предприятие
      tags:
      - app
    get:
      consumes:
      - application/json
      description: Получение информации о зарегистрированном сервере 1С.Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/AppServer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
      summary: Получение информации о зарегистрированном сервере 1С.Предприятие
      tags:
      - app
    post:
      consumes:
      - application/json
      description: Обновление информации о зарегистрированном сервере 1С.Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: app info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/AppServer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/AppServer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Обновление информации о зарегистрированном сервере 1С.Предприятие
      tags:
      - app
  /app/{app}/agent/admins:
    get:
      consumes:
      - application/json
      description: получение списка администраторов агента на сервере 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/serialize.UserInfo'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: получение списка администраторов агента на сервере 1С Предприятие
      tags:
      - admins
      - agent
    post:
      consumes:
      - application/json
      description: выполняет регистрацию нового адмнимистратор на агенте сервера 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: user info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/serialize.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.UserInfo'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: выполняет регистрацию нового адмнимистратор на агенте сервера 1С Предприятиеи
      tags:
      - admins
      - agent
  /app/{app}/agent/admins/{admin}:
    delete:
      consumes:
      - application/json
      description: Удаление администратора агента на сервере 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: admin name
        in: path
        name: admin
        required: true
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Удаление администратора агента на сервере 1С Предприятие
      tags:
      - admins
      - agent
...
swagger: "2.0"